Rearrange dependencies in the utilities.
